use CSS ends-with matcher
We test locally using npx http-server. In that context our content is
served at the root.

We publish globally using GitHub pages. And we have already told
GitHub to use another repository for our root path.

The ends-with matcher allows our code to work locally as well as work
when published.